NewHandsOn/study-session-coach/requirements.txt
gradio>=4.44.0
Pillow>=10.4.0
